refactor(components): add ApplicationApiUtils
